LY-84636 203126, There are identical sub-expressions 'pVerts[2] - pVerts[3]' ' to the left and to the right of the '^' operator.
The bug here is that the expression `pVerts[2] - pVerts[3]' ^ pVerts[2] - pVerts[3]'` has duplicated subexpressions either side of the cross-product operator. This is incorrect and will always evaluate to zero, it was probably the result of a typo at some point in this files long history.

I have made an educated guess at the fix for this but it will require some checking as this code is very maths heavy with zero comments and lots of operations going on. Based on my understanding of the intention here, the code is meant to be evaluating an aggregate cross product for two triangles in this quad. `(pVerts[1] - pVerts[0] ^ pVerts[2] - pVerts[0])` will find the perpendicular vector from pVert[0], and I have reckoned that the intention was to then find the perpendicular vector from pVert[2] which will be summed to get `n` which we can then use to create a quaternion for the physicalised geometry which will be added to the entity.